+++ src/main/java/org/getspout/vanilla/entity/projectile/Projectile.java
@@ -46,7 +46,7 @@ public class Projectile extends Controller {
 	}
 
 	@Override
-	public void onTick(float dt) {
+	public void onTick(long dt) {
 		Transform t = parent.getTransform();
 		//position += velocity.transform(rotation) * dt;
 		t.setPosition(t.getPosition().add(velocity.transform(t.getRotation()).scale(dt)));
